Add multiple sequence alignment to guideSet object

Calculates rough multiple sequence alignment and consensus sequence or imports from external fasta file

addAlignments(guideSet, files = NULL, max_gap_freq = 0.8,
  iterations = 2, refinements = 1, force = FALSE)

Arguments

files
data.frame. Imports multiple sequence alignments from file for provided families. For other families, the alignment is computed. Requires columns 'repname' and 'path' storing family identifiers and paths to the multiple sequence alignment, respectively.

max_gap_freq
Numeric between 0 and 1. Removes positions on alignment with higher gap (-) frequency than max_gap_freq.

iterations
Integer. Passed to AlignSeqs from DECIPHER package.

refinements
Integer. Passed to AlignSeqs from DECIPHER package.

force
Logical. If TRUE, overwrites existing alignments.

guideSet.
guideSet containing targets.

Value

Returns a guideSet object with multiple sequence alignment and consensus model.

See also

Examples

# NOT RUN {
gs <- createGuideSet(genome = Hsapiens, tes = te_annotation_df)
gs <- addTargets(gs, targets = c('THE1B', 'THE1C'))
# import multiple sequence alignment for THE1C and compute for THE1B
gs <- addAlignments(gs, files = data.frame('repname' = 'THE1C', path = 'path to multiple sequence alignment fasta file'))
gs@alignments
# }