Generates combinations QC plots

Computes multiple QC plots and arranges into a figure with subpanels:

a

Table showing total on/off-target scores (Son_tot and Soff_tot) for the best combination per number of guides (columns).

b

Total on- and off-target score per number of guides.

c

Total on- and off-target score per number of greedy search iterations. Colors indicate the number of guides (from yellow (at least 2) to black (defined by max_guides in the addCombinations function).

d

Total on- and off-target score for all computated combinations per number of guides.

e

Off-target coverage for the best guide combination per number of guides and off-target type (e.g. transposon family). Colors indicate predicted binding affinity to the off-target.

f

Distance to the nearest cis regulatory feature (defined by cis in the createGuideSet function) for the best combination number of guides.

g

On-target coverage for the best guide combination per number of guides and on-target transposon family. Colors indicate predicted binding affinity to the target.

h

Coverage for the best guide combination on the consensus model of transposon families. Colors indicate the number of guides (from yellow (1) to black (max_guides))

plotCombinations(guideSet)

Arguments

guideSet
guideSet containing combinations.

Value

list of ggplot objects.

See also

Examples

# NOT RUN {
gs <- createGuideSet(Hsapiens, tes = te_annotation_df)
gs <- addTargets(gs, targets = 'LTR13A')
gs <- plotTargets(gs)
# }